feat(intent): extracted intent into separate component. added initial svg lines
