+++ ordering/components/MemberIntentField.js
@@ -2,6 +2,8 @@ import React from 'react'
 import Slider from 'material-ui/Slider'
 import { merge } from 'ramda'
 
+import Intent from './Intent'
+
 export default MemberIntentField
 
 function MemberIntentField (props) {
@@ -36,63 +38,5 @@ function PreIntent (props) {
       desired: newValue,
       maximum: newValue
     })
-  }
-}
-
-/*
-  {renderSlider({ name: 'minimum', min: 0, max: value.desired })}
-  {renderSlider({ name: 'desired', min: value.minimum, max: value.maximum })}
-  {renderSlider({ name: 'maximum', max: 10, min: value.desired })}
-*/
-
-function Intent (props) {
-  const { onChange, value } = props
-
-  return (
-    <div>
-      {renderSlider({ name: 'minimum' })}
-      {renderSlider({ name: 'desired' })}
-      {renderSlider({ name: 'maximum' })}
-    </div>
-  )
-
-  function renderSlider (options) {
-    const { name } = options
-    return (
-      <Slider
-        onChange={handleChange(name)}
-        value={value[name]}
-        min={0}
-        max={10}
-        step={1}
-      />
-    )
-  }
-
-  function handleChange (key) {
-    return (ev, newValue) => {
-      var update = { [key]: newValue }
-
-      if (key == 'minimum') {
-        if (newValue > value.desired) {
-          update.desired = newValue
-        }
-        if (newValue > value.maximum) {
-          update.maximum = newValue
-        }
-      } else if (key == 'desired') {
-        const diff = newValue - value.desired
-        update.minimum = Math.max(0, value.minimum + diff)
-        update.maximum = Math.min(10, value.maximum + diff)
-      } else if (key == 'maximum') {
-        if (newValue < value.desired) {
-          update.desired = newValue
-        }
-        if (newValue < value.minimum) {
-          update.minimum = newValue
-        }
-      }
-      onChange(merge(value, update))
-    }
   }
 }
